requirement.txt
spacy==3.5.1
twitchio==2.5.0
pydub==0.25.1
openai==0.27.2
discord.py==2.2.3